pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrest-model-client"
version = "1.0.0"
description = "A simple, flexible Python HTTP client and API modeling toolkit built on httpx and pydantic."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Avi Zaguri" },]
keywords = ["rest", "development"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "colorlog>=6.10.1",
    "custom-python-logger>=2.0.10",
    "email-validator>=2.3.0",
    "httpx>=0.28.1",
    "pathlib>=1.0.1",
    "pre-commit>=4.5.0",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-plugins>=1.0.9",
    "python-base-toolkit>=1.0.1",
    "python-dotenv>=1.2.1",
    "python-simple-email-sender>=1.0.6",
    "python-vault>=0.0.3",
    "pyyaml>=6.0.3",
    "respx>=0.22.0",
    "setuptools>=80.9.0",
    "wheel>=0.45.1",
]

[project.urls]
Homepage = "https://github.com/aviz92/pyrest-model-client"
Repository = "https://github.com/aviz92/pyrest-model-client"
Issues = "https://github.com/aviz92/pyrest-model-client/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pyrest_model_client*"]
namespaces = false
